bunch of fixes for 2->3

+++ envi/qt/memwrite.py
@@ -269,7 +269,7 @@ class MemWriteWindow(QWidget):
 
 class MockEmu(object):
     def parseExpression(self, expr):
-        return long(eval(expr, {}, {}))
+        return int(eval(expr, {}, {}))
 
     def readMemory(self, va, size):
         return '\x90' * size
