+++ vtrace/platforms/posix.py
@@ -251,7 +251,5 @@ class PtraceMixin:
         for i in range(len(bytes)/wordsize):
             offset = wordsize*i
             dword = struct.unpack("L",bytes[offset:offset+wordsize])[0]
-            if ptrace(PT_WRITE_D, self.pid, long(address+offset), long(dword)) != 0:
+            if ptrace(PT_WRITE_D, self.pid, int(address+offset), int(dword)) != 0:
                 raise Exception("ERROR ptrace PT_WRITE_D failed!")
-
-
